+++ CHANGES.md
@@ -8,6 +8,7 @@ Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 * [#112] Added: `StreamEx.toArray(emptyArray)`
 * [#116] Added: `MoreCollectors.ifAllMatch()`
 * [#119] Added: `AbstractStreamEx.toImmutableList()`, `AbstractStreamEx.toImmutableSet()`, `EntryStream.toImmutableMap()`
+* [#123] Added: `StreamEx.into(collection)`, `EntryStream.into(map)`
 
 ### 0.6.2